+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 This file is for listing all the changes to this project
 
+## v0.1.1 BETA | May 14, 2023
+### Added
+- Adding the new Requirement for FancyMenu `is Internet Available`
+
+### Changed
+- Fixing the CFU Button offset in the fabric version
+
 ## v0.1.0 BETA | Dec 29, 2022
 - Port to 1.19.3